Add proper support for scion ctrl payloads in Go
Based on work by @shitz

Also:
- Ignore revocations receieved from local AS.
- Add some much-needed python logging re: rev infos.
- Changed common.Pld to have a WritePld method, making the interface not
  conflict with other possible Write methods.
- Fix an issue where revocations received from the local AS would be
  treated as an error instead of ignored.
- Added common.TypeOf to reduce the need for fmt.Sprintf("%T", v)
  everywhere.
- Removed a lot of unused/obsolete code from go/proto/.
- Change all proto ifid fields to be UInt64.
- Update to latest zombiezen.com/go/capnproto2, which grants us
  the ability to create canonical capnp messages from Go.

+++ proto/ifid.capnp
@@ -4,6 +4,6 @@ $Go.package("proto");
 $Go.import("github.com/netsec-ethz/scion/go/proto");
 
 struct IFID {
-   origIF @0 :UInt16;
-   relayIF @1 :UInt16;
+   origIF @0 :UInt64;
+   relayIF @1 :UInt64;
 }
